requirements.txt
scrapy<=2.4.1
SQLAlchemy<=1.3.22
kafka-python<=2.0.2
lz4<=3.1.1
crc32c<=2.2